piControl: Migrate to leds framework
The RevPi Core and RevPi Compact use different GPIO pins for power_red
(16 vs. 35) and a1_red (6 vs. 34).  Up until now we've hardcoded the
pins in project.h but that's not maintainable.  Toggling GPIO pins
directly was the wrong approach in the first place.

Going forward we'll declare LEDs and their associated GPIO pins in the
device tree, each with a trigger.  piControl in turn registers these
triggers and can thus toggle the LEDs independently of the underlying
hardware.

Use device-managed triggers to simplify their cleanup.

Signed-off-by: Lukas Wunner <lukas@wunner.de>