piControl: Introduce RevPi Compact process image
The process image of the RevPi Compact is the same as that of the
RevPi Core for compatibility.  The i8uStatus and RS-485 attributes
are meaningless on the Compact.

The layout of the digital I/O chips is identical to the DIO, likewise
for compatibility, however there are no counters or PWM values.
The upper half of the 16-bit "din" and "dout" attributes will always be
zero on the Compact, same for the bits in SDioModuleStatus pertaining to
channel 8-15.

The layout of the analog I/O chips is identical to the AIO for outgoing
data (two 16-bit values in mV).  Powerdown of the channels is not yet
defined, it could be controlled by a separate powerdown bit in the
process image, or by specifying a reserved value, say, 0xffff, or via a
PiCtory setting.

The layout of incoming data differs from the AIO:  There are eight
16-bit values (mV or °C/10).  Note that we're losing a portion of the
22-bit resolution, however the situation is the same on the AIO which
has a 24-bit ADC.  Users are always free to forgo piControl and access
the ADC directly via sysfs if they need the full resolution.  There is
also an aio_status byte which contains bits to convey SPI read or write
errors.  If the SPI read error bit is set, the ADC values may be
outdated or generally not trustworthy.  The AIO process image layout is
missing such a status byte, this appears to be an oversight.

Signed-off-by: Lukas Wunner <lukas@wunner.de>